Large changes to remote_files.js.
With RemoteFileManager.findPaths, it will be easy to associate remote
files with local files and it handles the special case where a file has
more than one parents.

Changes are persisted and parsed in a different way. Now pendingChanges
is a dictionary, where the key is file id and the value is updated
metadata. As a result, multiple changes to a single file are easily
consolidated and will be much easier to process.

Comment out findChildren, idChildrenMap for now. They're not used for
the time being.