Drive API: Improve the download method.
When options.etag is specified, the downloaded file is guaranteed to be
what is expected, no matter whether the file is changed between fetching
file's download URL and actual sending the download request.

When options.chunkCallback is specified, files larger than
DEFAULT_DOWNLOAD_CHUNK_SIZE will be downloaded in chunks to save memory.
chunkCallback is called when each chunk is received. It can return false
(by default undefined evaluated to false) and the next chunk will be
requested immediately. It can also return true (in this app, when the
previous chunk downloaded is not yet written to disk), indicating that
the next chunk will be explicitly requested by calling a callback
specified as the second parameter of chunkCallback.